Catch assets without file extensions
Based on the log file we got where it would hang ~97%
At the bottom was this:
2014-02-22 05:08:41,878 DEBUG      File "build\assets.py", line 322, in
index_data
2014-02-22 05:08:41,878 DEBUG    IndexError: list index out of range

line 322 was category = os.path.basename(asset[0]).split(".")[1]

This would return index out of bounds if there wasn't a period in the
key. Changed the line to splittext which will return a 2-tuple, index 0
is the filename index 1 is the "last extension"

Added the warning about assets that don't have extensions so modders can
fix them.